Move finder from pages/app/models to pages/lib.
