+++ pages/app/models/refinery/page.rb
@@ -2,7 +2,7 @@
 require 'friendly_id'
 require 'refinery/core/base_model'
 require 'refinery/pages/url'
-require 'refinery/page_finder'
+require 'refinery/pages/finder'
 
 module Refinery
   class Page < Core::BaseModel
@@ -79,14 +79,14 @@ module Refinery
 
       # Find page by path, checking for scoping rules
       def find_by_path(path)
-        PageFinder.find_by_path(path)
+        Pages::Finder.by_path(path)
       end
 
       # Helps to resolve the situation where you have a path and an id
       # and if the path is unfriendly then a different finder method is required
       # than find_by_path.
       def find_by_path_or_id(path, id)
-        PageFinder.find_by_path_or_id(path, id)
+        Pages::Finder.by_path_or_id(path, id)
       end
 
       # Helps to resolve the situation where you have a path and an id
@@ -107,7 +107,7 @@ module Refinery
       # pages table thus requiring us to find the attribute on the translations table
       # and then join to the pages table again to return the associated record.
       def by_title(title)
-        PageFinder.by_title(title)
+        Pages::Finder.by_title(title)
       end
 
       # Finds pages by their slug.  This method is necessary because pages
@@ -115,7 +115,7 @@ module Refinery
       # pages table thus requiring us to find the attribute on the translations table
       # and then join to the pages table again to return the associated record.
       def by_slug(slug, conditions={})
-        PageFinder.by_slug(slug, conditions)
+        Pages::Finder.by_slug(slug, conditions)
       end
 
       # Shows all pages with :show_in_menu set to true, but it also
@@ -133,7 +133,7 @@ module Refinery
 
       # Wrap up the logic of finding the pages based on the translations table.
       def with_globalize(conditions = {})
-        PageFinder.with_globalize(conditions)
+        Pages::Finder.with_globalize(conditions)
       end
 
       # Returns how many pages per page should there be when paginating pages
